requirements.txt
pandas
scikit-learn
textblob
streamlit
plotly
nltk
openai
schedule
python-dotenv
markdown
